Moved model checkpoints to a folder called checkpoints. Models (LSTM and CNN) are now in models instead of training script.
